+++ firebase.js
@@ -1,5 +1,5 @@
-const admin = require('firebase-admin');
-const serviceAccount = require('./config/hacka-35045-firebase-adminsdk-fbsvc-3081425674.json');
+const admin = require("firebase-admin");
+const serviceAccount = require("./fb_key.json"); //change to your own firebase service account key
 
 admin.initializeApp({
   credential: admin.credential.cert(serviceAccount)
